Move `InviteUserAsync` to `IInviteService`
